chore: updating code for selection controls for cleaner caption
