+++ packages/pluggableWidgets/selection-controls-web/src/components/CheckboxSelection/CheckboxSelection.tsx
@@ -1,6 +1,7 @@
 import classNames from "classnames";
 import { ReactElement, createElement } from "react";
 import { SelectionBaseProps, MultiSelector } from "../../helpers/types";
+import { CaptionContent } from "../CaptionContent";
 
 export function CheckboxSelection({
     selector,
@@ -42,7 +43,6 @@ export function CheckboxSelection({
                 aria-required={ariaRequired?.value}
             >
                 {options.map((optionId, index) => {
-                    const caption = selector.caption.get(optionId);
                     const isSelected = currentIds.includes(optionId);
                     const checkboxId = `${inputId}-checkbox-${index}`;
 
@@ -63,11 +63,7 @@ export function CheckboxSelection({
                                 onChange={e => handleChange(optionId, e.target.checked)}
                                 aria-describedby={`${inputId}-description`}
                             />
-                            <label htmlFor={checkboxId} className="widget-selection-controls-checkbox-label">
-                                {selector.customContentType === "no"
-                                    ? caption
-                                    : selector.caption.render(optionId, "options")}
-                            </label>
+                            <CaptionContent htmlFor={checkboxId}>{selector.caption.render(optionId)}</CaptionContent>
                         </div>
                     );
                 })}
@@ -75,26 +71,6 @@ export function CheckboxSelection({
                     <div className="widget-selection-controls-no-options">No options available</div>
                 )}
             </div>
-
-            {/* Clear all button */}
-            {/* {!isReadOnly && currentIds.length > 0 && (
-                <button
-                    type="button"
-                    className="widget-selection-controls-clear-all"
-                    onClick={() => selector.setValue([])}
-                    aria-label={a11yConfig.ariaLabels.clearSelection}
-                >
-                    Clear all selections
-                </button>
-            )} */}
-
-            {/* Accessibility status message */}
-            {/* <div id={`${inputId}-description`} className="sr-only" aria-live="polite" aria-atomic="true">
-                {currentIds.length > 0 &&
-                    `${a11yConfig.a11yStatusMessage.a11ySelectedValue} ${currentIds.map(id => selector.caption.get(id)).join(", ")}`}
-                {` ${a11yConfig.a11yStatusMessage.a11yOptionsAvailable} ${options.length}`}
-                {` ${a11yConfig.a11yStatusMessage.a11yInstructions}`}
-            </div> */}
         </div>
     );
 }
